+++ frontend/Login/comment.js
@@ -1,0 +1,10 @@
+import styles from "@/styles/comment.module.css";
+
+const Comment = () => {
+    return (
+        <p className={styles.comment}>
+            外部アカウントを使用して登録またはログインしますか？
+        </p>
+    )
+}
+export default Comment
