+++ frontend/src/styles/comment.module.css
@@ -1,0 +1,8 @@
+.comment {
+    width: 374px;
+    font-size: 14px;
+    left: 20px;
+    margin: 0 auto;
+    transform: translateY(320px);
+
+}
