--- frontend/src/styles/google-button.module.css
@@ -1,56 +1,0 @@
-/* GitHubButton.module.css */
-.Btn {
-    width: 45px;
-    height: 45px;
-    left: 80px;
-    top: 360px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: none;
-    background-color: transparent;
-    position: relative;
-    border-radius: 7px;
-    cursor: pointer;
-    transition: all .3s;
-}
-
-.svgContainer {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: white;
-    backdrop-filter: blur(0px);
-    letter-spacing: 0.8px;
-    border-radius: 10px;
-    transition: all .3s;
-    border: 1px solid rgba(156, 156, 156, 0.466);
-}
-
-.BG {
-    position: absolute;
-    content: "";
-    width: 100%;
-    height: 100%;
-    background: white;
-    z-index: -1;
-    border-radius: 10px;
-    pointer-events: none;
-    transition: all .3s;
-}
-
-.Btn:hover .BG {
-    transform: rotate(35deg);
-    transform-origin: bottom;
-}
-
-.Btn:hover .svgContainer {
-    background-color: rgba(156, 156, 156, 0.466);
-    backdrop-filter: blur(4px);
-}
-
-.svgContainer svg {
-    fill: white; /* アイコンの色を白に設定 */
-}
